requirements.txt
billiard==3.3.0.23
celery==3.1.24
certifi==2016.9.26
decorator==4.0.10
Django==1.8.2
django-celery==3.1.17
django-celery-results==1.0.1
django-common-helpers==0.9.0
django-cors-headers==1.2.2
django-cron==0.5.0
django-crontab==0.7.1
django-kombu==0.9.4
django-model-utils==2.6
djangorestframework==3.5.2
